requirements.txt
Flask
click
numpy
requests
progressbar2
flask_cors
azure-storage
pandas
xlrd